examples/pubsub/pyproject.toml
[tool.poetry]
name = "pubsub"
version = "0.1.0"
description = ""
authors = ["luiz zamboni"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-pubsub = "^2.23.0"
google-cloud-storage = "^2.18.2"
faker = "^30.8.1"
pytz = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
